Jacob Contests and Skills

+++ EliteAPI/Models/DTOs/Outgoing/Skills.cs
@@ -3,5 +3,5 @@
 public class SkillDto
 {
     public required string Type { get; set; }
-    public long Exp { get; set; } = 0;
+    public double Exp { get; set; } = 0;
 }
